Abstract lockpick puzzle into component
